latest from coda-oss/main

+++ externals/coda-oss/modules/c++/io/source/PipeStream.cpp
@@ -59,7 +59,7 @@ sys::SSize_T io::PipeStream::readln(sys::byte *cStr,
     while (!feof(pipe))
     {
         // get the next line or return null
-        if (fgets(cStr, strLenPlusNullByte, pipe) != NULL)
+        if (fgets(cStr, static_cast<int>(strLenPlusNullByte), pipe) != NULL)
         {
             return strlen(cStr);
         }
